+++ interfaces/chats_controller.py
@@ -7,7 +7,7 @@ from PyQt5 import QtWidgets
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QImage, QPixmap
 
-from .raw_interfaces.contact_card import Ui_ContactCard
+from .raw.contact_card import Ui_ContactCard
 
 
 class ChatItem(Ui_ContactCard, QtWidgets.QWidget):
