+++ main.py
@@ -4,7 +4,7 @@ import sass
 
 from PyQt5.QtWidgets import QMainWindow
 
-from interfaces.raw_interfaces.main_window import Ui_MainWindow
+from interfaces.raw.main_window import Ui_MainWindow
 from interfaces.chats_controller import ChatsController
 from interfaces.message_controller import MessageController
